xd
xd

+++ tarea4/prueba2/programa2.c
@@ -5,6 +5,7 @@
 #include <sys/ipc.h>
 #include <signal.h>
 #include <unistd.h>
+#include <errno.h>
 
 #define KEY 1234
 
@@ -17,24 +18,38 @@ struct message {
 void *thread_function(void *arg) {
     int msqid = *(int*)arg;
     struct message msg;
-    long mtype = (long)pthread_self() % 2 + 1;
+    long mtype = (long)arg == 1 ? 1 : 2;  // Thread 1 maneja tipo 1, Thread 2 maneja tipo 2
 
     while(1) {
         if (msgrcv(msqid, &msg, sizeof(msg) - sizeof(long), mtype, 0) < 0) {
             perror("msgrcv");
-            exit(1);
+            continue;
         }
 
         printf("Thread de tipo %ld recibió: Señal %d para PID %d\n", mtype, msg.signal, msg.pid);
 
         if (kill(msg.pid, 0) == 0) {
-            if (kill(msg.pid, msg.signal) < 0) {
-                perror("kill");
+            if (msg.signal == SIGSTOP || msg.signal == SIGCONT) {
+                // Para SIGSTOP y SIGCONT, usamos kill() directamente
+                if (kill(msg.pid, msg.signal) < 0) {
+                    perror("kill (SIGSTOP/SIGCONT)");
+                } else {
+                    printf("Señal %d enviada al proceso %d\n", msg.signal, msg.pid);
+                }
             } else {
-                printf("Señal %d enviada al proceso %d\n", msg.signal, msg.pid);
+                // Para otras señales, usamos pthread_kill() para mayor compatibilidad
+                if (pthread_kill(msg.pid, msg.signal) != 0) {
+                    perror("pthread_kill");
+                } else {
+                    printf("Señal %d enviada al proceso %d\n", msg.signal, msg.pid);
+                }
             }
         } else {
-            printf("Proceso con PID %d no existe o ya ha terminado.\n", msg.pid);
+            if (errno == ESRCH) {
+                printf("Proceso con PID %d no existe.\n", msg.pid);
+            } else {
+                perror("kill");
+            }
         }
     }
     return NULL;
@@ -43,14 +58,16 @@ void *thread_function(void *arg) {
 int main() {
     int msqid;
     pthread_t thread1, thread2;
+    int thread1_type = 1;
+    int thread2_type = 2;
 
     if ((msqid = msgget(KEY, 0666)) < 0) {
         perror("msgget");
         exit(1);
     }
 
-    if (pthread_create(&thread1, NULL, thread_function, &msqid) != 0 ||
-        pthread_create(&thread2, NULL, thread_function, &msqid) != 0) {
+    if (pthread_create(&thread1, NULL, thread_function, &thread1_type) != 0 ||
+        pthread_create(&thread2, NULL, thread_function, &thread2_type) != 0) {
         perror("pthread_create");
         exit(1);
     }
